remade shader loading

+++ include/primitives.hpp
@@ -5,8 +5,8 @@
 class Primitive
 {
 public:
-	static float* triangle(int x1, int y1, int z1, int x2, int y2, int z2, int x3, int y3, int z3);
-	static float* rectangle(int x1, int y1, int z1, int x2, int y2, int z2, int x3, int y3, int z3, int x4, int y4, int z4);
+	static float* triangle(float x1, float y1, float z1, float x2, float y2, float z2, float x3, float y3, float z3);
+	static float* rectangle(float x1, float y1, float z1, float x2, float y2, float z2, float x3, float y3, float z3, float x4, float y4, float z4);
 private:
 };
