+++ src/primitives.cpp
@@ -1,6 +1,6 @@
 #include <primitives.hpp>
 
-float* Primitive::triangle(int x1, int y1, int z1, int x2, int y2, int z2, int x3, int y3, int z3) {
+float* Primitive::triangle(float x1, float y1, float z1, float x2, float y2, float z2, float x3, float y3, float z3) {
 	float triangle[] = {
 		x1, y1, z1,
 		x2, y2, z2,
@@ -9,7 +9,7 @@ float* Primitive::triangle(int x1, int y1, int z1, int x2, int y2, int z2, int x
 	return triangle;
 }
 
-float* Primitive::rectangle(int x1, int y1, int z1, int x2, int y2, int z2, int x3, int y3, int z3, int x4, int y4, int z4) {
+float* Primitive::rectangle(float x1, float y1, float z1, float x2, float y2, float z2, float x3, float y3, float z3, float x4, float y4, float z4) {
 	float rectangle[] = {
 		x1, y1, z1,
 		x2, y2, z2,
